reorganize marker warning reporters
